cloud_common: schemas: revert service_config renaming
This reverts commit d4820d2f68d799383e7c6d0c4cd2350f5ab77e67.
